refactor(backend): refactor code to avoid redundant
