+++ src/backend/app/app/tests/api/test_users.py
@@ -3,7 +3,6 @@ from sqlalchemy.orm import Session
 from typing import Dict
 
 from app.crud.user import search_user_by_username
-from app.crud.tag import search_tag
 from app.tests.utils.utils import random_string, random_email
 from app.core.settings import settings
 
@@ -158,36 +157,3 @@ def test_update_current_user_password(client: TestClient, db: Session, regular_u
         "new_password": "passwd",
     })
     assert r.status_code == 200
-
-def test_create_user_tag(client: TestClient, db: Session, superuser_token_headers: Dict[str, str]) -> None:
-    r = client.post('usertag/create', headers=superuser_token_headers, json={
-        "name": random_string(k=10),
-    })
-
-    assert r.status_code == 200
-    tag = r.json()
-    assert search_tag(session=db, name=tag["name"])
-
-def test_create_existing_tag(client: TestClient, superuser_token_headers: Dict[str, str]) -> None:
-    tag_name = random_string(k=10)
-    r = client.post('usertag/create', headers=superuser_token_headers, json={
-        "name": tag_name,
-    })
-
-    assert r.status_code == 200
-
-    r = client.post('usertag/create', headers=superuser_token_headers, json={
-        "name": tag_name,
-    })
-
-    assert r.status_code == 400
-    response = r.json()
-    assert response["detail"]=="There is alredy a tag with this name"
-
-def test_create_tag_regular_user(client: TestClient, regular_user_token_headers: Dict[str, str]) -> None:
-    r = client.post('usertag/create', headers=regular_user_token_headers, json={
-        "name": random_string(),
-    })
-    assert r.status_code == 400
-    response = r.json()
-    assert response["detail"]=='User must be a superuser'
